// apps/nexus/src/apis/AnimeScheduleApi/dataSource/dtos/AnimeScheduleApiDtos.ts
export type AiringSchedule = {
	title: string;
	romaji: string;
	english: string;
	native: string;
	route: string;
	delayedFrom: string;
	delayedUntil: string;
	status: string;
	episodeDate: string;
	episodeNumber: number;
	episodes: number;
	lengthMin: number;
	imageVersionRoute: string;
	airingStatus: string;
};